requirements.txt
requests>=2.25.1
python-dotenv>=0.19.0
pytest>=7.0.1
tqdm>=4.62.3